requirements.txt
Django
djangorestframework
django-allauth
django-cors-headers
"psycopg[binary]"
requests
django-environ
pyJWT
cryptography
drf-spectacular
python-dateutil
kafka-python
celery
redis
black
isort
flake8
mypy
django-stubs